fix the narrowing fix 😁(8c6e342e004ae6799b34675af99f741535e78d46)

+++ inc/inject.h
@@ -3,7 +3,7 @@
 #include "defs.h"
 #include <elf.h>
 
-int inject_section(cliArgs *args);
+long inject_section(cliArgs *args);
 void modify_section_header(Elf64_Addr addr, sectionHeader *section, long offset);
 char *get_section_name(int index);
 sectionHeader *get_section_by_name(char *section_name);
